Added simple, though working obstacle generation Moved Obstacle manager to a separate file Fixed car speed issues (car was 'moving' at speed 0)
